Refactor code and enhance JSON serialization
In this commit, several significant changes were made across multiple files. In `Authorization.cs`, the `Uri` property was renamed to `Url`. This change was reflected in `EventService.cs` where the `uri` variable was updated to use the new `Url` property.

Additionally, a `JsonSerializerOptions` object was created in `EventService.cs` to customize the JSON serialization process. The `EventMeta.cs` and `EventResult.cs` files were updated to include the `System.Runtime.Serialization` namespace and decorated with `XmlInclude` and `KnownType` attributes for `PoaCreateUnifiedRequest`.

The namespace in `EventResult.cs` was changed from `ParusRx.Astral.API.Models` to `ParusRx.Astral.Domain`. A new class `EventResultJsonConverter` was added in `EventResultJsonConverter.cs` to provide custom serialization and deserialization logic for the `EventResult` class.

Lastly, in `PoaCreateUnifiedRequest.cs`, several properties in the `AuthorizedRepresentative` and `Authority` records were changed from required to optional, allowing these properties to be null.