Update Dockerfile and .NET publish settings
Updated the Dockerfile to change the `WORKDIR` to `/src` and use `base` as the final image with `/app` as the working directory. Added `dotnet publish` command to publish the `Ips.IdentityProvider.API.csproj` project with `UseAppHost` set to `false`. The contents of `/app/publish` from the `publish` image are now copied into the current directory. The `HEALTHCHECK` command has been updated to use `wget` for health checks. Added `ENTRYPOINT` command to run `Ips.IdentityProvider.API.dll` with `dotnet` on container start.