Updated Docker settings and downgraded health check package
Updated `docker-compose.yml` to run the application in a production environment. The base image in `Dockerfile` was updated to use the newer version of the .NET runtime. The `HEALTHCHECK` command in `Dockerfile` was uncommented for Docker to periodically check the application's health. Downgraded the `AspNetCore.HealthChecks.UI.Client` package in `Ips.IdentityProvider.API.csproj` to an older version. Added the `HealthChecks.UI.Client` namespace in `Program.cs` for easier access to its classes and methods.