Change `Code` property type to nullable int
In the `ParusRx.Egisz.Frmr.Domain` namespace, specifically within the `MpSpec.cs` and `Spec.cs` files, the type of the `Code` property in both the `MpSpec` and `Spec` records has been updated from a nullable string (`string?`) to a nullable integer (`int?`). This modification alters the representation and handling of the `Code` property across these domain models, with potential implications for serialization, deserialization, and related logic.