Add TO_INFORM to TemplateKey enum and fix ROUTE syntax
The changes made to the `TemplateKey` enum in `TemplateKey.cs` include the addition of a comma after the `ROUTE` enum value, which is necessary for proper syntax when adding new enum values. Additionally, a new enum value `TO_INFORM` was introduced, complete with an XML serialization attribute `[XmlEnum("TO_INFORM")]` and a descriptive summary comment `/// The inform.`.