Add new paths to monitored directories in mis-api.yml
The changes in the `mis-api.yml` file include the addition of new paths to the list of monitored directories for both the `push` and `pull_request` events. Specifically, the paths `src/Services/Egisz/Frmo/Frmo/**` and `src/Services/Egisz/Frmo/Frmo.Domain/**` have been added. This means that any changes in these directories will now trigger the workflows defined in this YAML file.